+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.1
+
+**Fixed**
+
+* The `domContentLoaded` function no longer executes its callback more than once
+
 ## 3.1.0
 
 **Added**